User: Split shared utils to modules
