+++ client/lib/user/shared-utils/raw-current-user-fetch.js
@@ -1,0 +1,10 @@
+/**
+ * Internal dependencies
+ */
+import wpcom from 'calypso/lib/wp';
+
+export function rawCurrentUserFetch() {
+	return wpcom.me().get( {
+		meta: 'flags',
+	} );
+}
